requirements.txt
apispec==5.2.1
Flask==2.1.2
Flask-Cors==3.0.10
flask-smorest==0.37.0
Flask-SocketIO==5.1.2
marshmallow==3.15.0
python-certifi-win32==1.6
python-engineio==4.3.2
python-jwt==2.0.1
requests
requests-toolbelt==0.7.0
Werkzeug==2.1.2
coloredlogs==15.0.1
python-dotenv==0.20.0
suds==1.1.2
setuptools==58.1.0
gunicorn==20.1.0
pymongo
jmespath==0.10.0
python-dateutil==2.8.1
six==1.15.0
urllib3==1.26.3
PyJWT==2.3.0